requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
streamlit>=1.36.0
numpy>=1.26.0
faiss-cpu>=1.13.0
transformers>=4.45.0
